test: update backpatch tests
